# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Green Flag Award Foundation API",
    "version": "0.0.0",
    "description": "Slice 3 registration API plus identity/RBAC/audit foundation endpoints."
  },
  "paths": {
...
    "/api/v1/registrations": {
      "post": {
        "operationId": "submitRegistration",
        "summary": "Submit a park registration with eligibility and duplicate checks.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationSubmissionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Registration accepted or eligibility failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationSubmissionResponse"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate acknowledgement required.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registrationId}": {
      "get": {
        "operationId": "getRegistration",
        "summary": "Return a registration summary.",
        "parameters": [
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Registration summary.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationSummary"
                }
              }
            }
          },
          "404": {
            "description": "Registration not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registrationId}/location-lookup": {
      "post": {
        "operationId": "resolveRegistrationLocation",
        "summary": "Return mock W3W, OS, and ONS location suggestions requiring applicant confirmation.",
        "parameters": [
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationLocationLookupRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Location suggestion.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationLocationSuggestion"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/registrations/{registrationId}/verify-email": {
      "post": {
        "operationId": "verifyRegistrationEmail",
        "summary": "Verify registration email token.",
        "parameters": [
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailVerificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerificationLandingResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/registration-review-queue": {
      "get": {
        "operationId": "listRegistrationReviews",
        "summary": "List verified registrations awaiting admin review.",
        "responses": {
          "200": {
            "description": "Registration review queue.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminRegistrationReviewQueueResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/registration-review-queue/{registrationId}/approve": {
      "post": {
        "operationId": "approveRegistration",
        "summary": "Approve a verified registration.",
        "parameters": [
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRegistrationDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Park activation response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParkActivationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Invalid transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/registration-review-queue/{registrationId}/reject": {
      "post": {
        "operationId": "rejectRegistration",
        "summary": "Reject a verified registration.",
        "parameters": [
          {
            "name": "registrationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminRegistrationDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Park activation response.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ParkActivationResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Invalid transition.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            }
          }
        }
      },
      "RegistrationEligibilityInput": {
        "type": "object",
        "required": [
          "publiclyAccessible",
          "freeToEnter",
          "minimumSizeConfirmed"
        ],
        "properties": {
          "publiclyAccessible": {
            "type": "boolean"
          },
          "freeToEnter": {
            "type": "boolean"
          },
          "minimumSizeConfirmed": {
            "type": "boolean"
          }
        }
      },
      "RegistrationEligibilityResult": {
        "type": "object",
        "required": [
          "eligible",
          "failedCriteria"
        ],
        "properties": {
          "eligible": {
            "type": "boolean"
          },
          "failedCriteria": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "publicly_accessible",
                "free_to_enter",
                "minimum_size"
              ]
            }
          }
        }
      },
      "DuplicateWarning": {
        "type": "object",
        "required": [
          "hasPotentialDuplicate",
          "matchedFields",
          "acknowledged"
        ],
        "properties": {
          "hasPotentialDuplicate": {
            "type": "boolean"
          },
          "matchedFields": {
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "park_name",
                "postcode",
                "address"
              ]
            }
          },
          "acknowledged": {
            "type": "boolean"
          }
        }
      },
      "RegistrationLocationLookupRequest": {
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "properties": {
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "postcode": {
            "type": "string"
          },
          "w3wAddress": {
            "type": "string"
          }
        }
      },
      "RegistrationLocationSuggestion": {
        "type": "object",
        "required": [
          "source",
          "label",
          "latitude",
          "longitude",
          "w3wAddress",
          "requiresApplicantConfirmation"
        ],
        "properties": {
          "source": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "latitude": {
            "type": "number"
          },
          "longitude": {
            "type": "number"
          },
          "w3wAddress": {
            "type": "string"
          },
          "parkNameSuggestion": {
            "type": "string"
          },
          "sizeBand": {
            "type": "string"
          },
          "localAuthority": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "constituency": {
            "type": "string"
          },
          "requiresApplicantConfirmation": {
            "type": "boolean",
            "const": true
          }
        }
      },
      "RegistrationSubmissionRequest": {
        "type": "object",
        "required": [
          "parkName",
          "organisationName",
          "contactName",
          "contactEmail",
          "addressLine1",
          "town",
          "country",
          "eligibility",
          "location"
        ],
        "properties": {
          "parkName": {
            "type": "string"
          },
          "organisationName": {
            "type": "string"
          },
          "contactName": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string",
            "format": "email"
          },
          "addressLine1": {
            "type": "string"
          },
          "town": {
            "type": "string"
          },
          "postcode": {
            "type": "string"
          },
          "country": {
            "type": "string"
          },
          "eligibility": {
            "$ref": "#/components/schemas/RegistrationEligibilityInput"
          },
          "duplicateAcknowledged": {
            "type": "boolean"
          },
          "location": {
            "$ref": "#/components/schemas/RegistrationLocationLookupRequest"
          }
        }
      },
      "RegistrationSubmissionResponse": {
        "type": "object",
        "required": [
          "registrationId",
          "status",
          "eligibility",
          "duplicateWarning",
          "verificationRequired",
          "notificationIntents"
        ],
        "properties": {
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "eligibility": {
            "$ref": "#/components/schemas/RegistrationEligibilityResult"
          },
          "duplicateWarning": {
            "$ref": "#/components/schemas/DuplicateWarning"
          },
          "verificationRequired": {
            "type": "boolean"
          },
          "notificationIntents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "RegistrationSummary": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RegistrationSubmissionResponse"
          },
          {
            "type": "object",
            "required": [
              "parkName",
              "organisationName",
              "contactEmail",
              "submittedAt"
            ],
            "properties": {
              "parkName": {
                "type": "string"
              },
              "organisationName": {
                "type": "string"
              },
              "contactEmail": {
                "type": "string",
                "format": "email"
              },
              "submittedAt": {
                "type": "string",
                "format": "date-time"
              }
            }
          }
        ]
      },
      "EmailVerificationRequest": {
        "type": "object",
        "required": [
          "token"
        ],
        "properties": {
          "token": {
            "type": "string"
          }
        }
      },
      "VerificationLandingResponse": {
        "type": "object",
        "required": [
          "registrationId",
          "status",
          "emailVerified",
          "nextStep"
        ],
        "properties": {
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "nextStep": {
            "type": "string"
          }
        }
      },
      "AdminRegistrationReviewQueueResponse": {
        "type": "object",
        "required": [
          "items"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "AdminRegistrationDecisionRequest": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
      "ParkActivationResponse": {
        "type": "object",
        "required": [
          "registrationId",
          "registrationStatus",
          "parkStatus",
          "notificationIntents"
        ],
        "properties": {
          "registrationId": {
            "type": "string",
            "format": "uuid"
          },
          "registrationStatus": {
            "type": "string"
          },
          "parkId": {
            "type": "string",
            "format": "uuid"
          },
          "parkStatus": {
            "type": "string"
          },
          "notificationIntents": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    }
  }
}